refactor: change structure for jwt, bands, generes,tracks, users modules
