+++ src/modules/allSchemas.ts
@@ -1,11 +1,11 @@
 import { artistSchema } from './artists/artist.schema.js';
 import { albumSchema } from './albums/album.schema.js';
-import { bandSchema } from './bands/schema.js';
-import { genreSchema } from './genres/schema.js';
+import { bandSchema } from './bands/band.schema.js';
+import { genreSchema } from './genres/genre.schema.js';
 import { favouriteSchema } from './favourites/schema.js';
-import { userSchema } from './users/schema.js';
-import { trackSchema } from './tracks/schema.js';
-import { jwtSchema } from './jwt/schema.js';
+import { userSchema } from './users/user.schema.js';
+import { trackSchema } from './tracks/track.schema.js';
+import { jwtSchema } from './jwt/jwt.schema.js';
 
 export const allSchemas = `
 ${artistSchema}
