+++ src/modules/bands/band.schema.ts
