+++ src/modules/bands/resolvers/band.queries.ts
@@ -1,4 +1,4 @@
-export const bandsResolver = {
+export const bandsQueriesResolver = {
   bands: async (_, __, { dataSources }) => {
     return dataSources.bandsService.getAllBands();
   },
