+++ src/modules/jwt/resolvers/jwt.queries.ts
@@ -1,4 +1,4 @@
-export const jwtResolver = {
+export const jwtQueriesResolver = {
   login: async (_, { email, password }, { dataSources }) => {
     return dataSources.jwtService.login(email, password);
   },
