+++ src/modules/users/resolvers/users.queries.ts
@@ -1,4 +1,4 @@
-export const usersResolver = {
+export const usersQueriesResolver = {
   user: async (_, { id }, { dataSources }) => {
     return dataSources.usersService.getUser(id);
   },
